PBDSegment: Convert to a simpler abstract class
Make PBDSegment a simple abstrct class with just the bare minimum in it. Move
most method implementations to PBDRegularSegment.
